[ALPS04792966] DISP: fix ion bug
[Detail]
ion_phys() param 3 should be 64 bit.
disp only use low 32 bit.

MTK-Commit-Id: 2393d388671f54de7794c226b533872e4d2960b4

CR-Id: ALPS04792966
Feature: [Module]Display Driver

+++ drivers/misc/mediatek/video/mt6785/videox/primary_display.c
@@ -1995,7 +1995,7 @@ static int sec_buf_ion_alloc(int buf_size)
 {
 #ifdef MTK_FB_ION_SUPPORT
 	size_t mva_size = 0;
-	unsigned long int sec_hnd = 0;
+	unsigned int sec_hnd = 0;
 	/* ion_phys_addr_t sec_hnd = 0; */
 	unsigned long align = 0; /* 4096 alignment */
 	struct ion_mm_data mm_data;
